- refactor logger to own namespace
